Rename package from ai.jetbrains.code to ai.koog.
