+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/tools/GeographyQueryTool.kt
@@ -1,4 +1,4 @@
-package ai.jetbrains.code.integration.tests.tools
+package ai.koog.integration.tests.tools
 
 import ai.koog.agents.core.tools.SimpleTool
 import ai.koog.agents.core.tools.Tool
@@ -7,35 +7,35 @@ import ai.koog.agents.core.tools.ToolParameterDescriptor
 import ai.koog.agents.core.tools.ToolParameterType
 import kotlinx.serialization.Serializable
 
-object GenericParameterTool : SimpleTool<GenericParameterTool.Args>() {
+object GeographyQueryTool : SimpleTool<GeographyQueryTool.Args>() {
     @Serializable
     data class Args(
-        val requiredArg: String,
-        val optionalArg: String? = null
+        val query: String,
+        val language: String? = null
     ) : Tool.Args
 
     override val argsSerializer = Args.serializer()
 
     override val descriptor = ToolDescriptor(
-        name = "generic_parameter_tool",
-        description = "A tool that demonstrates handling of required and optional parameters",
+        name = "geography_query_tool",
+        description = "A tool for retrieving geographical information such as capitals of countries",
         requiredParameters = listOf(
             ToolParameterDescriptor(
-                name = "requiredArg",
-                description = "A required string parameter that must be provided",
+                name = "query",
+                description = "The geographical query (e.g., 'capital of France')",
                 type = ToolParameterType.String
             )
         ),
         optionalParameters = listOf(
             ToolParameterDescriptor(
-                name = "optionalArg",
-                description = "An optional string parameter that can be omitted",
+                name = "language",
+                description = "The language code to return the response in (e.g., 'en', 'fr')",
                 type = ToolParameterType.String
             )
         )
     )
 
     override suspend fun doExecute(args: Args): String {
-        return "Generic parameter tool executed with requiredArg: ${args.requiredArg}, optionalArg: ${args.optionalArg ?: "not provided"}"
+        return "Geography query processed: ${args.query}, language: ${args.language ?: "not specified"}"
     }
 }
