+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/utils/TestUtils.kt
@@ -1,4 +1,4 @@
-package ai.jetbrains.code.integration.tests.utils
+package ai.koog.integration.tests.utils
 
 import ai.koog.agents.core.tools.*
 import kotlinx.coroutines.flow.Flow
